requirements.txt
fastapi==0.115.12
matplotlib==3.10.1
natsort==8.4.0
numpy==1.26.4
opencv-python==4.11.0.86
Pillow==11.1.0
python-multipart==0.0.20
seaborn==0.13.2
tqdm==4.67.1
uvicorn==0.34.0